Apply proper name to Fakturant static library
Applies a proper name to the static library of Fakturant.
CMake silently adds "lib" prefix to all the static libraries, unless a
special PREFIX property is overridden for a target. This resulted in
building a static library with name "liblib_Fakturant.a".
This commit fixes this by setting PREFIX to an empty string. This
results in a build with the intended name, "lib_Fakturant.a".